Update check_update_local.py
I ran into some issues with check_update_local.py on Red Hat distros, so I fixed it, and thought I would share with you the fix. Please excuse my rather clumsy fix as this is maybe the fourth piece of python code I have had to hack/fix. 

The popen cmd process would hang for yum check-updates for security if the output of the cmd was too long (in my case from packages/repos being excluded). So I added a -q to both the commands, which fixes the issue. But from what I could figure out the popen should probably use the communicate() method.  

Additionally there were several updates with RHEL/CentOS/AmazonLinux sources that did not end with "updates", so I converted it to a tuple (lines 203,204) and added the unique strings that the sources I found had .

I should add that this fix does not work with Amazon Linux, nor does the original. This is because although Amazon Linux is built on RHEL, it uses /etc/os-release for version information instead of /etc/redhat-release. I am working this as well and hopefully will have an update soon